Create PriorityQueue.cpp
This is the main file for a priority queue (a priority queue is similar to the line at a hospital - the most severe case gets served first, even if it's the newest number to enter the queue). In this case, the numbers are randomly generated and given priority based on their size, but this can be reprogrammed to give priority based on other factors.